chore: move story md utils to a seperate file
